+++ .prettierrc.js
@@ -1,0 +1,16 @@
+/** @type {import('prettier').Options} */
+module.exports = {
+  printWidth: 80,
+  endOfLine: 'lf',
+  tabWidth: 2,
+  useTabs: false,
+  semi: true,
+  quoteProps: 'as-needed',
+  singleQuote: true,
+  jsxSingleQuote: true,
+  trailingComma: 'none',
+  bracketSameLine: false,
+  bracketSpacing: true,
+  arrowParens: 'avoid',
+  proseWrap: 'preserve'
+};