+++ src/fetch.ts
@@ -1,12 +1,12 @@
 import nodeFetch, { RequestInit, Response } from 'node-fetch';
 
-import { ubiAppId } from './auth';
+import { IUbiAuth, ubiAppId } from './auth';
 
 const promiseTimeout = <T>(promise: Promise<T>, ms: number, reject = true) =>
   Promise.race([promise, new Promise((res, rej) => setTimeout(() => reject ? rej : res, ms))]);
 
 export default <T>(url: string, options: Partial<RequestInit> = {}) =>
-  async (token?: string): Promise<T> => {
+  async (token?: string, auth?: IUbiAuth): Promise<T> => {
 
     const { headers, ...optionsRest } = options;
 
@@ -19,7 +19,10 @@ export default <T>(url: string, options: Partial<RequestInit> = {}) =>
             'Content-Type': 'application/json; charset=UTF-8',
             'Ubi-AppId': ubiAppId,
             ...token && { 'Authorization': token },
-            ...headers && { ...headers }
+            ...headers && { ...headers },
+            ...auth && {
+              'Ubi-SessionId': auth.sessionId
+            }
           }
         },
         ...optionsRest && { ...optionsRest }